requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
backports.entry-points-selectable==1.1.0
cattrs==1.8.0
certifi==2021.5.30
cfgv==3.3.1
charset-normalizer==2.0.4
distlib==0.3.2
Faker==8.12.1
filelock==3.0.12
flake8==3.9.2
identify==2.2.14
idna==3.2
iniconfig==1.1.1
mccabe==0.6.1
nodeenv==1.6.0
packaging==21.0
password==0.2
platformdirs==2.3.0
pluggy==1.0.0
pre-commit==2.15.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
username==2.1.0
virtualenv==20.7.2
